+++ internal/filter.go
@@ -1,7 +1,7 @@
 // filter.go takes care of filtering the desired informations
 // from Packages for each architecture.
 
-package src
+package internal
 
 import (
 	"bufio"
