+++ internal/packages.go
@@ -1,4 +1,4 @@
-package src
+package internal
 
 import (
 	"io"
